feat: adds support for context key/values
if --lookForContextData is specified, the tool will use --contextPrefix and/or --contextSuffix to search for contextual key/value pairs for each language/key. For example, if a --contextPrefix is passed with '_context:' and '--keys help-button', the tool will search for another key/value pair in the reference file with key '_context:help-button'. The value of this, if found, will be passed along to the AI provider to give give additional context to the translation request.